match function data to framework

+++ include/awnn/common.h
@@ -19,8 +19,8 @@ extern "C" {
 #endif
 
 typedef unsigned int uint;
-typedef float T;
-//typedef double T;
+//typedef float T;
+typedef double T;
 
 typedef unsigned int label_t;
 label_t *label_make_random(uint nr_elem, uint range);
